added codepath to create VKInstance from external one
