+++ cmake/settings.cmake
@@ -61,21 +61,4 @@ else()
   option(WISDOM_USE_FMT "Build Wisdom with fmtlib" OFF)
 endif()
 
-if(WISDOM_modules)
-  if(CMAKE_MINOR_VERSION LESS 27)
-    message(
-      WARNING
-        "WARNING: modules are at beta support, use this build mode at your own risk."
-    )
-  endif()
-  if(CMAKE_MINOR_VERSION EQUAL 25)
-    set(CMAKE_EXPERIMENTAL_CXX_MODULE_CMAKE_API
-        "3c375311-a3c9-4396-a187-3227ef642046")
-  else()
-    set(CMAKE_EXPERIMENTAL_CXX_MODULE_CMAKE_API
-        "2182bf5c-ef0d-489a-91da-49dbc3090d2a")
-  endif()
-  set(CMAKE_EXPERIMENTAL_CXX_MODULE_DYNDEP 1)
-endif()
-
 set(WISDOM_VERSION "v${PROJECT_VERSION_MAJOR}_${PROJECT_VERSION_MINOR}")
