+++ examples/custom/lut_windowed/CMakeLists.txt
@@ -29,7 +29,7 @@ add_executable(
 wis_install_deps(${PROJECT_NAME})
 
 target_link_libraries(
-  ${PROJECT_NAME} PRIVATE wis-header-only wisdom-platform-headers glm::glm Wil)
+  ${PROJECT_NAME} PRIVATE wis::wisdom-vk-headers wis::wisdom-platform glm::glm Wil)
 set_target_properties(${PROJECT_NAME} PROPERTIES CXX_STANDARD 20)
 target_include_directories(${PROJECT_NAME} PRIVATE ${fpng_SOURCE_DIR}/src)
