+++ wisdom/include/wisdom/vulkan/vk_device.h
@@ -13,6 +13,30 @@
 #include <wisdom/generated/vulkan/vk_structs.hpp>
 
 namespace wis {
+constexpr static inline std::array required_extensions{
+    VK_KHR_SWAPCHAIN_EXTENSION_NAME, // for Swapchain
+    VK_KHR_TIMELINE_SEMAPHORE_EXTENSION_NAME, // for Fence
+    VK_KHR_SYNCHRONIZATION_2_EXTENSION_NAME, // for barriers
+
+    VK_KHR_DEDICATED_ALLOCATION_EXTENSION_NAME, // for Allocator
+    VK_KHR_GET_PHYSICAL_DEVICE_PROPERTIES_2_EXTENSION_NAME, // for Allocator
+    VK_KHR_BIND_MEMORY_2_EXTENSION_NAME, // for Allocator
+    VK_KHR_MAINTENANCE_4_EXTENSION_NAME, // for Allocator
+
+    VK_EXT_DESCRIPTOR_BUFFER_EXTENSION_NAME, // for PushDescriptor
+    VK_EXT_EXTENDED_DYNAMIC_STATE_2_EXTENSION_NAME, // for Tessellation control point count
+    VK_KHR_DYNAMIC_RENDERING_EXTENSION_NAME, // for dynamic render pass
+
+    VK_VALVE_MUTABLE_DESCRIPTOR_TYPE_EXTENSION_NAME, // for Mutable Descriptor Type
+    VK_EXT_MUTABLE_DESCRIPTOR_TYPE_EXTENSION_NAME,
+
+    VK_EXT_CUSTOM_BORDER_COLOR_EXTENSION_NAME, // for Border Color
+
+    // VK_KHR_RAY_TRACING_PIPELINE_EXTENSION_NAME,
+    // VK_KHR_RAY_QUERY_EXTENSION_NAME,
+    // VK_KHR_ACCELERATION_STRUCTURE_EXTENSION_NAME,
+};
+
 struct InternalFeatures {
     bool has_descriptor_buffer : 1 = false;
     bool push_descriptor_bufferless : 1 = false;
@@ -149,6 +173,9 @@ private:
 
 [[nodiscard]] WIS_INLINE wis::ResultValue<wis::VKDevice>
 VKCreateDevice(wis::VKAdapter in_adapter) noexcept;
+
+[[nodiscard]] WIS_INLINE wis::ResultValue<wis::VKDevice>
+VKCreateDeviceEx(wis::VKAdapter in_adapter, std::span<const char*> add_extensions) noexcept;
 
 } // namespace wis
